Extract JavadocElementParser logic to it's own utility class
