+++ openapi-maven-plugin/src/main/java/io/github/kbuntrock/javadoc/JavadocWrapper.java
@@ -2,24 +2,13 @@ package io.github.kbuntrock.javadoc;
 
 import com.github.javaparser.javadoc.Javadoc;
 import com.github.javaparser.javadoc.JavadocBlockTag;
-import com.github.javaparser.javadoc.description.JavadocDescription;
-import com.github.javaparser.javadoc.description.JavadocDescriptionElement;
-import com.github.javaparser.javadoc.description.JavadocInlineTag;
-import com.github.javaparser.javadoc.description.JavadocSnippet;
 import io.github.kbuntrock.utils.Logger;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
-import java.util.function.Function;
-import java.util.function.Predicate;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 /**
  * @author Kevin Buntrock
@@ -81,17 +70,11 @@ public class JavadocWrapper {
 	}
 
 	public Optional<String> getSummary() {
-		return processJavadocElements(
-			elements -> elements
-				.filter(onlyTagsOfType(JavadocTag.SUMMARY))
-				.map(JavadocWrapper::toTagContent));
+		return JavadocElementParser.getSummary(javadoc.getDescription(), endOfLineReplacement);
 	}
 
 	public Optional<String> getDescription() {
-		return processJavadocElements(
-			elements -> elements
-				.filter(JavadocWrapper::onlySnippetsAndFormattingTags)
-				.map(JavadocWrapper::formatJavadocText));
+		return JavadocElementParser.getDescription(javadoc.getDescription(), endOfLineReplacement);
 	}
 
 	public boolean isInheritTagFound() {
@@ -105,74 +88,14 @@ public class JavadocWrapper {
 				Logger.INSTANCE.getLogger().debug(entry.getKey() + " : " + entry.getValue().getContent().toText());
 			}
 		}
-
 	}
 
 	public void printReturn() {
 		if(blockTagsByType != null) {
 			final Optional<JavadocBlockTag> returnTag = getReturnBlockTag();
-			if(returnTag.isPresent()) {
-				Logger.INSTANCE.getLogger().debug("Return : " + returnTag.get().getContent().toText());
-			}
+			returnTag.ifPresent(javadocBlockTag ->
+				Logger.INSTANCE.getLogger().debug("Return : " + javadocBlockTag.getContent().toText()));
 		}
 	}
 
-	private static String formatJavadocText(JavadocDescriptionElement javadocElement) {
-		if(javadocElement instanceof JavadocInlineTag) {
-			JavadocInlineTag tag = (JavadocInlineTag) javadocElement;
-			switch(JavadocTag.fromString(tag.getName())) {
-				case CODE:
-					return "`" + tag.getContent().trim() + "`";
-				case SEE:
-					return "*" + tag.getContent().trim() + "*";
-				case LINK:
-					Pattern p = Pattern.compile("href=\"(.*?)\"");
-					Matcher m = p.matcher(tag.getContent());
-					String url = null;
-					if (m.find()) {
-						url = m.group(1);
-					}
-					return "[" + url + "](" + url + ")";
-				default:
-					return tag.getContent();
-			}
-		}
-
-		return javadocElement.toText();
-	}
-
-	private Optional<String> processJavadocElements(Function<Stream<JavadocDescriptionElement>, Stream<String>> elementProcessor) {
-		return Optional.ofNullable(javadoc.getDescription())
-			.map(JavadocDescription::getElements)
-			.map(Collection::stream)
-			.map(elementProcessor)
-			.map(s -> s.collect(Collectors.joining("\n")))
-			.map(JavadocWrapper::removeNewlinesIfActivated)
-			.filter(text -> !text.isEmpty())
-			.map(String::trim);
-	}
-
-	private static Predicate<JavadocDescriptionElement> onlyTagsOfType(JavadocTag tag) {
-		return e -> (e instanceof JavadocInlineTag && tag.toString().toLowerCase().equals(((JavadocInlineTag) e).getName()));
-	}
-
-	private static boolean onlySnippetsAndFormattingTags(JavadocDescriptionElement e) {
-		return onlySnippets(e) || (
-			e instanceof JavadocInlineTag &&
-				JavadocTag.isFormattingTag(((JavadocInlineTag) e).getName()));
-	}
-
-	private static boolean onlySnippets(JavadocDescriptionElement e) {
-		return e instanceof JavadocSnippet;
-	}
-
-	private static String toTagContent(JavadocDescriptionElement t) {
-		return ((JavadocInlineTag) t).getContent().trim();
-	}
-
-	private static String removeNewlinesIfActivated(String text) {
-		return endOfLineReplacement != null
-			? text.replaceAll("\\r\\n", endOfLineReplacement).replaceAll("\\n", endOfLineReplacement)
-			: text;
-	}
 }
